+++ game/weather.py
@@ -10,7 +10,7 @@ from typing import Optional, TYPE_CHECKING, Any
 from dcs.cloud_presets import Clouds as PydcsClouds
 from dcs.weather import CloudPreset, Weather as PydcsWeather, Wind
 
-from game.settings import Settings
+from game.settings.settings import Settings, NightMissions
 from game.utils import Distance, Heading, meters, interpolate, Pressure, inches_hg
 
 from game.theater.seasonalconditions import determine_season
@@ -301,7 +301,10 @@ class Conditions:
         settings: Settings,
     ) -> Conditions:
         _start_time = cls.generate_start_time(
-            theater, day, time_of_day, settings.night_disabled
+            theater,
+            day,
+            time_of_day,
+            settings.night_day_missions,
         )
         return cls(
             time_of_day=time_of_day,
@@ -315,15 +318,23 @@ class Conditions:
         theater: ConflictTheater,
         day: datetime.date,
         time_of_day: TimeOfDay,
-        night_disabled: bool,
+        night_day_missions: NightMissions,
     ) -> datetime.datetime:
-        if night_disabled:
+        if night_day_missions == NightMissions.OnlyDay:
             logging.info("Skip Night mission due to user settings")
             time_range = {
                 TimeOfDay.Dawn: (8, 9),
                 TimeOfDay.Day: (10, 12),
                 TimeOfDay.Dusk: (12, 14),
                 TimeOfDay.Night: (14, 17),
+            }[time_of_day]
+        elif night_day_missions == NightMissions.OnlyNight:
+            logging.info("Skip Day mission due to user settings")
+            time_range = {
+                TimeOfDay.Dawn: (0, 3),
+                TimeOfDay.Day: (3, 6),
+                TimeOfDay.Dusk: (21, 22),
+                TimeOfDay.Night: (22, 23),
             }[time_of_day]
         else:
             time_range = theater.daytime_map[time_of_day.value]